requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
ortools>=9.15.0
pandas>=2.2.0
openpyxl>=3.1.0
python-multipart>=0.0.17
jinja2>=3.1.0
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Development tools
black>=24.0.0
ruff>=0.1.0
pytest>=8.0.0

